+++ app/app.go
@@ -68,6 +68,7 @@ import (
 	"github.com/kiichain/kiichain/v7/app/keepers"
 	"github.com/kiichain/kiichain/v7/app/upgrades"
 	v7_3_1 "github.com/kiichain/kiichain/v7/app/upgrades/v7_3_1"
+	v8_0 "github.com/kiichain/kiichain/v7/app/upgrades/v8_0"
 	"github.com/kiichain/kiichain/v7/client/docs"
 )
 
@@ -78,6 +79,7 @@ var (
 	// Upgrades is a list of all the upgrades that are available for the application.
 	Upgrades = []upgrades.Upgrade{
 		v7_3_1.Upgrade,
+		v8_0.Upgrade,
 	}
 )
 
@@ -329,6 +331,7 @@ func (app *KiichainApp) setAnteHandler(txConfig client.TxConfig, maxGasWanted ui
 		EvmKeeper:              app.EVMKeeper,
 		FeeAbstractionKeeper:   app.FeeAbstractionKeeper,
 		FeegrantKeeper:         app.FeeGrantKeeper,
+		CircuitKeeper:          &app.CircuitKeeper,
 		IBCKeeper:              app.IBCKeeper,
 		FeeMarketKeeper:        app.FeeMarketKeeper,
 		SignModeHandler:        txConfig.SignModeHandler(),